Moved UserFactory classes into entity_factories
Hopefully this doesn't cause any unwanted code conflicts
