+++ src/main/java/entity_factories/UserFactory.java
@@ -1,4 +1,6 @@
-package entity_layer;
+package entity_factories;
+
+import entity_layer.User;
 
 public interface UserFactory {
     User create(String username, String email, String password);
